Added Some functionality and fixes to string library(havent checked them yet) updated window library, math library(unfinished) updated defines, started creating test library, also some refactoring

+++ include/sklc_lib/vector.h
@@ -1,26 +1,26 @@
 #ifndef _SKLC_LIB_VECTOR_H_
 #define _SKLC_LIB_VECTOR_H_
 
-#include <stdbool.h>
+#include <sklc_lib/defines.h>
 
 // Used by the defines, DONT'USE IT BY YOURSELF
 typedef struct _vector { 
 
-    // WARNING: DO NOT MANUALLY CHANGE ANYTHING BELOW!!!
     void* data;                 // Vector's data
 
-    unsigned long long stride;  // Size of the element
-    unsigned long long length;  // Number of elements currently used
-    unsigned long long capacity;// How much items can it hold before reallocating;
+    // WARNING: DO NOT MANUALLY CHANGE ANYTHING BELOW!!!
+    u64 stride;  // Size of the element
+    u64 length;  // Number of elements currently used
+    u64 capacity;// How much items can it hold before reallocating;
 
     void (*push_back)(struct _vector* vec, const void* data);
     void* (*pop_back)(struct _vector* vec);
 
-    void (*set_at)(struct _vector* vec, unsigned long long index, void* val);
-    void*(*get_at)(struct _vector* vec, unsigned long long index);
+    void (*set_at)(struct _vector* vec, u64 index, void* val);
+    void*(*get_at)(struct _vector* vec, u64 index);
 
-    void (*insert_at)(struct _vector* vec, const void* data, unsigned long long index);
-    void*(*remove_at)(struct _vector* vec, unsigned long long index);
+    void (*insert_at)(struct _vector* vec, const void* data, u64 index);
+    void*(*remove_at)(struct _vector* vec, u64 index);
 
     void (*clear)(struct _vector* vec);
 
@@ -28,7 +28,7 @@ typedef struct _vector {
 
     // adds passed length to the current capacity
     
-    void (*reserve)(struct _vector* vec, unsigned long long capacity);
+    void (*reserve)(struct _vector* vec, u64 capacity);
 
     bool is_struct_ptr;
     
@@ -36,8 +36,10 @@ typedef struct _vector {
 
 #define vector _vector
 
-vector _vector_create(unsigned long long capacity, unsigned long long stride);
-vector* _vector_create_ptr(unsigned long long capacity, unsigned long long stride);
+vector _vector_create(u64 capacity, u64 stride);
+
+// Use thes ONLY when you want to create an allocated vector or an array of vectors
+vector* _vector_create_ptr(u64 capacity, u64 stride);
 void vector_destroy(vector* vec);
 
 #define vector_create(type) _vector_create(1, sizeof(type))
